requirements.txt
numpy==2.1.3
pandas==2.3.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
tensorflow==2.19.0
matplotlib==3.10.3
pillow==11.3.0
fastapi==0.116.1
python-jose==3.5.0
onnxruntime==1.15.1
passlib[bcrypt]==1.7.4
openpyxl==3.1.5
pydantic[email]==2.2.0
python-multipart==0.0.20
dask==2025.7.0
pyarrow==21.0.0
onnxruntime==1.3.0
plotly==6.3.0
tf2onnx==1.16.1
openai==1.99.9
python-dotenv==1.1.1